Extracted AppBuilder to separate module.
